[fidi]: Implement a /healthz mechanism
Have fidi return a HTTP_OK response on the request to /healthz. Since
the healthz request need not contain a valid request, other
meachnisms shall be implemented to establish when the application is
in an unhealthy state. This state should be changeable by a request,
temprarily or permanently, in the future.

Signed-off-by: Manoj Srivastava <srivasta@google.com>
Change-Id: I3a9acf795853053e607fdbf805dd473d6031c05b