[fidi]: Enable requests disabling fidi for a specified time
This commit enables a request to ask the fidi instance to be totally
unresponsive for a specified time period, dropping all requests on the
floor, including health checks (By dropping requests, i this context,
we mean that no response is sent; we still listen on the port, and
something looks at each request. Eventually, the request will time out
with a message about an empty server response). This can be used to
simulate a restart.

Signed-off-by: Manoj Srivastava <srivasta@google.com>
Change-Id: I0cd54826429ef7ddb8859d17dc6ed398cbd6f62f